requirements.txt
Flask==2.3.3
Flask-Mail==0.10.0
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
playwright==1.40.0